Move CookbookManifest to its own file
